Add missing err check.  Add missing table field.
When iterating and scanning over a slice of rows, you must check for
errors return by the rows.Scan() function.  If you don't, the iteration
will carry on happily and fill your objects with default empty values
silently.

The commend id column was missing from the query for comments on a post,
or comments of a user.  It was silently left as is due to the
non-checking of the error code returned by rows.Scan(), which once
enabled, clearly mentionned that I was scanning for 7 fields but only
getting 6 in my SQL query.